Main

OTRecod.compute_average_error_bound — Function

compute_average_error_bound(path; norme=1)

Compute a lower bound on the best average prediction error that one can obtain with a specific type of data sets path: path of the directory containing the data set

OTRecod.empirical_estimator — Function

empirical_estimator(path; observed = [])

Get an empirical estimator of the distribution of Z conditional to Y and X on base A and reciprocally on base B obtain with a specific type of data sets

path: path of the directory containing the data set

observed: if nonempty, list of indices of the observed covariates; this allows to exclude some latent variables.

OTRecod.test_ncds — Method

test_ncds(file_name)

Run every method on an input file generated from the real NCDS dataset

OTRecod.test_params_group — Function

test_params_group(bench_path, nbfiles::Int64=1, norme::Int64=0)

Run the relaxed group transport with a range of relaxation parameters

bench_path: name of the benchmark directory

nbfiles: number of files considered per directory of the bench, 0 for all

norme : 0, 1 or 2, norm used for distances in the space of covariates

OTRecod.test_params_joint — Function

test_params_joint(bench_path, nbfiles::Int64=1, norme::Int64=0)

Run the transport of covariates and outcomes with a range of parameters for relaxation and regularization

bench_path: name of the benchmark directory

nbfiles: number of files considered per directory of the bench, 0 for all

norme : 0, 1 or 2, norm used for distances in the space of covariates