Joint

Joint

OTRecod.OT_joint — Function.

Model where we directly compute the distribution of the outcomes for each individual or for sets of indviduals that similar values of covariates aggregatetol: quantify how much individuals' covariates must be close for aggregation regnorm: norm1, norm2 or entropy depending on the type of regularization percentclosest: percent of closest neighbors taken into consideration in regularization lambdareg: coefficient measuing the importance of the regularization term fulldisp: if true, write the transported value of each individual; otherwise, juste write the number of missed transports solverdisp: if false, do not display the outputs of the solver

source