Plots

OTRecod.plot_benchmark — Function

plot_benchmark(datapath,outputpath,ymin=0.0, ymax=0.8)

Boxplot of each simulation with all methods

datapath: path of the folder where the data files are stored

outputpath: path of the folder where the result outputs are stored

ymin: minimum ordinate value

ymax: maximum ordinate value

OTRecod.plot_scenario — Function

plot_scenario(outputpath, scenario, method, ymin=0.0, ymax=0.4)

Boxplot comparing several scenarios solved with one given method

outputpath: path of the folder where the result outputs are stored

scenario: beginning of the names of all the scenarios that are to be compared

method: name of the solution method; should be a string with following format "methodname-maxrelax-lambdareg", e.g., "joint-0.4-0.1"

ymin: minimum ordinate value

ymax: maximum ordinate value