Utilities

Helper functions

OTRecod.Instance — Type.

Instance(data_file, norme)

Definition and initialization of an Instance structure

datafile : file name

norme : integer

source

OTRecod.Solution — Type.

Solution(t, jointYZA, jointYZB)

Solution(t, jointYZA, jointYZB, estimatorZA, estimatorYB)

tsolve : solution time

jointYZA : joint distribution of Y and Z in A

jointYZB : joint distribution of Y and Z in B

estimatorZA : estimator of probability of Z for individuals in base A

estimatorYB : estimator of probability of Y for individuals in base B

source

OTRecod.average_distance_to_closest — Method.

average_distance_to_closest(instance, percent_closest)

Compute the cost between pairs of outcomes as the average distance between covariations of individuals with these outcomes, but considering only the percent closest neighbors

source

OTRecod.avg_distance_closest — Method.

avg_distance_closest(instance, database1, database2, outcome, m1, m2, 
                     percent_closest)

Compute the average distance between individuals of base1 with modality m1 for outcome and individuals of base2 with modality m2 for outcome Consider only the percent_closest individuals in the computation of the distance

source

OTRecod.bound_prediction_error — Function.

bound_prediction_error(instance, norme, aggregate_tol)

Compute a bound on the average prediction error in each base. The bound is computed as the expected prediction error assuming that the distribution of Z in base A (and that of Y in base B) is known, and the prediction done with the value that maximizes the probability

source

OTRecod.compute_distrib_error — Method.

compute_distrib_error(insttance, solution, empiricalZA, empiricalYB)

Compute errors in the conditional distributions of a solution

source

OTRecod.compute_pred_error — Function.

Compute prediction errors in a solution

source

OTRecod.disp_inst_info — Method.

disp_inst_info(instance)

Display information about the distance between the modalities

source

OTRecod.empirical_distribution — Function.

empirical_distribution(instance, norme, aggregate_tol)

Return the empirical cardinality of the joint occurrences of (C=x,Y=mA,Z=mB) in both bases

source