Documentation

OTRecod.jl

Documentation for OTRecod.jl

Installation

In a Julia session switch to pkg> mode to add NPSMC:

julia>] # switch to pkg> mode
pkg> add https://github.com/otrecoding/OTRecod.jl

Alternatively, you can achieve the above using the Pkg API:

julia> using Pkg
julia> pkg"add https://github.com/otrecoding/OTRecod.jl"

When finished, make sure that you're back to the Julian prompt (julia>) and bring OTRecod into scope:

julia> using OTRecod

You can test the package with

julia>] # switch to pkg> mode
pkg> test OTRecod

To run an example from a dataset

OTRecod.run_directory — Function.

run_directory(path, method; outname="result.out", 
                            maxrelax=0.0, 
                            lambda_reg=0.0, 
                            nbfiles=0, 
                            norme=0, 
                            percent_closest=0.2)

Run one given method on a given number of data files of a given directory The data files must be the only files with extension ".txt" in the directory

path : name of the directory

method : :group or :joint

nbfiles: number of files considered, 0 if all the data files are tested

norme : 1 or 2, norm used for distances in the space of covariates

(see runallmethods for the description of other parameters)

source

julia> using OTRecod