OTRecod.jl

Documentation for OTRecod.jl

Installation

In a Julia session switch to pkg> mode to add NPSMC:

julia>] # switch to pkg> mode
pkg> add https://github.com/otrecoding/OTRecod.jl

Alternatively, you can achieve the above using the Pkg API:

julia> using Pkg
julia> pkg"add https://github.com/otrecoding/OTRecod.jl"

When finished, make sure that you're back to the Julian prompt (julia>) and bring OTRecod into scope:

julia> using OTRecod

You can test the package with

julia>] # switch to pkg> mode
pkg> test OTRecod

To run an example from a dataset

OTRecod.run_directory — Function

run_directory(path, method)
run_directory(path, method, outname)
run_directory(path, method, outname, maxrelax)
run_directory(path, method, outname, maxrelax, lambda_reg)
run_directory(
    path,
    method,
    outname,
    maxrelax,
    lambda_reg,
    nbfiles
)
run_directory(
    path,
    method,
    outname,
    maxrelax,
    lambda_reg,
    nbfiles,
    norme
)
run_directory(
    path,
    method,
    outname,
    maxrelax,
    lambda_reg,
    nbfiles,
    norme,
    percent_closest
)
run_directory(
    path,
    method,
    outname,
    maxrelax,
    lambda_reg,
    nbfiles,
    norme,
    percent_closest,
    observed
)

Run one given method on a given number of data files of a given directory The data files must be the only files with extension ".txt" in the directory

path : name of the directory

method : :group or :joint

maxrelax: maximum percentage of deviation from expected probability masses

lambda_reg: coefficient measuring the importance of the regularization term

nbfiles: number of files considered, 0 if all the data files are tested

norme : 0, 1 or 2, norm used for distances in the space of covariates

percent_closest: percent of closest neighbors taken in the computation of the costs (both distance and regularization related)

observed: if nonempty, list of indices of the observed covariates; this allows to exclude some latent variables.

julia> using OTRecod

Instance

OTRecod.Instance — Type

struct Instance

Definition and initialization of an Instance structure

datafile : file name

norme : ( 1 : Cityblock, 2 : Euclidean, 3 : Hamming )

indXA : indexes of subjects of A with given X value

indXB : indexes of subjects of B with given X value

Solution

OTRecod.Solution — Type

mutable struct Solution

tsolve : solution time

jointYZA : joint distribution of Y and Z in A

jointYZB : joint distribution of Y and Z in B

estimatorZA : estimator of probability of Z for individuals in base A

estimatorYB : estimator of probability of Y for individuals in base B