Group

Group

OTRecod.ot_group — Function.

ot_group(inst, percent_closest=0.2, maxrelax=0.0, norme=0, 
         indiv_method=sequential, full_disp=false, solver_disp=false)

Model of group transport percent_closest: percent of closest neighbors taken in the computation of the costs

maxrelax: maximum percentage of deviation from expected probability masses

indiv_method: specifies the method used to get individual transport from group joint probabilities

full_disp: if true, write the transported value of each individual; otherwise, juste write the number of missed transports

solver_disp: if false, do not display the outputs of the solver

source

OTRecod.individual_from_group_closest — Function.

individual_from_group_closest(inst, jointprobaA, jointprobaB, percent_closest=1.0)

Sequentially assign the modality of the individuals to that of the closest neighbor in the other base until the joint probability values are met

source

OTRecod.individual_from_group_optimal — Function.

individual_from_group_optimal(inst, jointprobaA,
     jointprobaB; percent_closest=1.0)

Solve an optimization problem to get the individual transport that minimizes total distance while satisfying the joint probability computed by the model by group

source