Joint

OTRecod.ot_joint — Function

ot_joint(inst)
ot_joint(inst, maxrelax)
ot_joint(inst, maxrelax, lambda_reg)
ot_joint(inst, maxrelax, lambda_reg, percent_closest)
ot_joint(inst, maxrelax, lambda_reg, percent_closest, norme)
ot_joint(
    inst,
    maxrelax,
    lambda_reg,
    percent_closest,
    norme,
    aggregate_tol
)
ot_joint(
    inst,
    maxrelax,
    lambda_reg,
    percent_closest,
    norme,
    aggregate_tol,
    full_disp
)
ot_joint(
    inst,
    maxrelax,
    lambda_reg,
    percent_closest,
    norme,
    aggregate_tol,
    full_disp,
    solver_disp
)

Model where we directly compute the distribution of the outcomes for each individual or for sets of indviduals that similar values of covariates

aggregate_tol: quantify how much individuals' covariates must be close for aggregation

reg_norm: norm1, norm2 or entropy depending on the type of regularization

percent_closest: percent of closest neighbors taken into consideration in regularization

lambda_reg: coefficient measuring the importance of the regularization term

full_disp: if true, write the transported value of each individual; otherwise, juste write the number of missed transports

solver_disp: if false, do not display the outputs of the solver