Helper functions

OTRecod.average_distance_to_closest — Method

average_distance_to_closest(instance, percent_closest)

Compute the cost between pairs of outcomes as the average distance between covariations of individuals with these outcomes, but considering only the percent closest neighbors

OTRecod.avg_distance_closest — Method

avg_distance_closest(instance, database1, database2,

outcome, m1, m2, percent_closest)

Compute the average distance between individuals of base1 with modality m1 for outcome and individuals of base2 with modality m2 for outcome

Consider only the percent_closest individuals in the computation of the distance

OTRecod.bound_prediction_error — Function

bound_prediction_error(instance, norme, aggregate_tol)

Compute a bound on the average prediction error in each base. The bound is computed as the expected prediction error assuming that the distribution of Z in base A (and that of Y in base B) is known, and the prediction done with the value that maximizes the probability

OTRecod.compute_distrib_error! — Method

compute_distrib_error!(solution, instance, empiricalZA, empiricalYB)

Compute errors in the conditional distributions of a solution

OTRecod.compute_distrib_error — Method

compute_distrib_error(instance, solution, empiricalZA, empiricalYB)

Compute errors in the conditional distributions of a solution

OTRecod.compute_pred_error — Function

compute_pred_error(inst, sol, proba_disp, mis_disp, full_disp)

Compute prediction errors in a solution

OTRecod.compute_pred_error! — Function

compute_pred_error(instance, sol, proba_disp=false,
                   mis_disp=false, full_disp=false)

Compute prediction errors in a solution

OTRecod.disp_inst_info — Method

disp_inst_info(instance)

Display information about the distance between the modalities

OTRecod.empirical_distribution — Function

empirical_distribution(instance, norme, aggregate_tol)

Return the empirical cardinality of the joint occurrences of (C=x,Y=mA,Z=mB) in both bases