Helper functions

OTRecod.aggregate_per_covar_mixed — Function

aggregate_per_covar_mixed(inst)
aggregate_per_covar_mixed(inst, norme)
aggregate_per_covar_mixed(inst, norme, aggregate_tol)

OTRecod.average_distance_to_closest — Method

average_distance_to_closest(inst, percent_closest)

Compute the cost between pairs of outcomes as the average distance between covariations of individuals with these outcomes, but considering only the percent closest neighbors

OTRecod.avg_distance_closest — Method

avg_distance_closest(
    inst,
    base1,
    base2,
    outcome,
    m1,
    m2,
    percent_closest
)

Compute the average distance between individuals of base1 with modality m1 for outcome and individuals of base2 with modality m2 for outcome

Consider only the percent_closest individuals in the computation of the distance

OTRecod.bound_prediction_error — Function

bound_prediction_error(inst)
bound_prediction_error(inst, norme)
bound_prediction_error(inst, norme, aggregate_tol)

Compute a bound on the average prediction error in each base. The bound is computed as the expected prediction error assuming that the distribution of Z in base A (and that of Y in base B) is known, and the prediction done with the value that maximizes the probability

OTRecod.compute_distrib_error! — Method

compute_distrib_error!(sol, inst, empiricalZA, empiricalYB)

Compute errors in the conditional distributions of a solution

OTRecod.compute_distrib_error — Method

compute_distrib_error(inst, sol, empiricalZA, empiricalYB)

Compute errors in the conditional distributions of a solution

OTRecod.compute_distrib_error_3covar — Method

compute_distrib_error_3covar(
    sol,
    inst,
    empiricalZA,
    empiricalYB
)

OTRecod.compute_pred_error — Function

compute_pred_error(inst, sol)
compute_pred_error(inst, sol, proba_disp)
compute_pred_error(inst, sol, proba_disp, mis_disp)
compute_pred_error(
    inst,
    sol,
    proba_disp,
    mis_disp,
    full_disp
)

Compute prediction errors in a solution

OTRecod.compute_pred_error! — Function

compute_pred_error!(sol, inst)
compute_pred_error!(sol, inst, proba_disp)
compute_pred_error!(sol, inst, proba_disp, mis_disp)
compute_pred_error!(
    sol,
    inst,
    proba_disp,
    mis_disp,
    full_disp
)

Compute prediction errors in a solution

OTRecod.disp_inst_info — Method

disp_inst_info(inst)

Display information about the distance between the modalities

OTRecod.empirical_distribution — Function

empirical_distribution(inst)
empirical_distribution(inst, norme)
empirical_distribution(inst, norme, aggregate_tol)

Return the empirical cardinality of the joint occurrences of (C=x,Y=mA,Z=mB) in both bases